+++ src/listeners/client/messageDelete.js
@@ -1,7 +1,11 @@
 const { Listener } = require('@eartharoid/dbf');
-const { AuditLogEvent } = require('discord.js');
+const {
+	AuditLogEvent, MessageFlags,
+} = require('discord.js');
 const { logMessageEvent } = require('../../lib/logging');
-const { quick } = require('../../lib/threads');
+const { pools } = require('../../lib/threads');
+
+const { crypto } = pools;
 
 module.exports = class extends Listener {
 	constructor(client, options) {
@@ -43,7 +47,7 @@ module.exports = class extends Listener {
 				const archived = await client.prisma.archivedMessage.findUnique({ where: { id: message.id } });
 				if (archived?.content) {
 					if (!content) {
-						const string = await quick('crypto', worker => worker.decrypt(archived.content));
+						const string = await crypto.queue(w => w.decrypt(archived.content));
 						content = JSON.parse(string).content; // won't be cleaned
 					}
 				}
@@ -73,7 +77,7 @@ module.exports = class extends Listener {
 			}
 		}
 
-		if (message.author.id !== client.user.id && !message.flags.has('Ephemeral')) {
+		if (message.author.id !== client.user.id && !message.flags.has(MessageFlags.Ephemeral)) {
 			await logMessageEvent(this.client, {
 				action: 'delete',
 				diff: {