+++ src/routes/api/admin/guilds/[guild]/export.js
@@ -1,17 +1,34 @@
-const {
-	spawn,
-	Pool,
-	Worker,
-} = require('threads');
 const { Readable } = require('node:stream');
-const { cpus } = require('node:os');
 const archiver = require('archiver');
 const { iconURL } = require('../../../../../lib/misc');
 const pkg = require('../../../../../../package.json');
+const { pools } = require('../../../../../lib/threads');
 
-// a single persistent pool shared across all exports
-const poolSize = Math.ceil(cpus().length / 4); // ! ceiL: at least 1
-const pool = Pool(() => spawn(new Worker('../../../../../lib/workers/export.js')), { size: poolSize });
+const { export: pool } = pools;
+
+/**
+ * Tracks currently running exports to prevent spamming exports
+ */
+const exportsRunning = {};
+const exportTasks = {};
+
+/**
+ * Release a ticket export lock
+ * @param id
+ */
+function releaseExport(id) {
+	delete exportsRunning[id];
+	const tasks = exportTasks[id];
+	// cancel all still running tasks to prevent clogging up threads for an already aborted request
+	if (tasks && tasks.length > 0) {
+		tasks.forEach(task => {
+			try {
+				task.cancel();
+			}catch (e){ /* empty */ }
+		});
+	}
+	delete exportTasks[id];
+}
 
 module.exports.get = fastify => ({
 	/**
@@ -28,6 +45,24 @@ module.exports.get = fastify => ({
 
 		client.log.info(`${member.user.username} requested an export of "${guild.name}"`);
 
+		// Check if an export is already running for this guild
+		if (Object.keys(exportsRunning).includes(id)) {
+			const time = exportsRunning[id];
+			// Check if a minute has already passed - something probably failed but prevented this guild to be removed from the list
+			if (time + 60000 <= (new Date().getDate())) {
+				exportsRunning[id] = new Date().getTime();
+			} else {
+				return res.status(429).send('An export is already running. Please wait for it to finish and try again afterwards.');
+			}
+		} else {
+			exportsRunning[id] = new Date().getTime();
+		}
+
+		// Detect if this request is aborted or closed and stop the export threads
+		req.raw.on('close', () => {
+			releaseExport(id);
+		});
+
 		// TODO: sign so the importer can ensure files haven't been added (important for attachments)
 		const archive = archiver('zip', {
 			comment: JSON.stringify({
@@ -40,11 +75,15 @@ module.exports.get = fastify => ({
 		});
 
 		archive.on('warning', err => {
-			if (err.code === 'ENOENT') client.log.warn(err);
-			else throw err;
+			if (err.code === 'ENOENT') {
+				client.log.warn(err);
+			} else {
+				throw err;
+			}
 		});
 
 		archive.on('error', err => {
+			releaseExport(id);
 			throw err;
 		});
 
@@ -68,7 +107,9 @@ module.exports.get = fastify => ({
 			return t;
 		});
 
+		// V1
 		const ticketsStream = Readable.from(ticketsGenerator());
+
 		async function* ticketsGenerator() {
 			try {
 				let done = false;
@@ -86,6 +127,11 @@ module.exports.get = fastify => ({
 					take,
 					where: { guildId: id },
 				};
+				// create worker index array
+				if (!exportTasks[id]) {
+					exportTasks[id] = [];
+				}
+
 				do {
 					const batch = await client.prisma.ticket.findMany(findOptions);
 					if (batch.length < take) {
@@ -95,7 +141,11 @@ module.exports.get = fastify => ({
 						findOptions.cursor = { id: batch[take - 1].id };
 					}
 					// ! map (parallel) not for...of (serial)
-					yield* batch.map(async ticket => (await pool.queue(worker => worker.exportTicket(ticket)) + '\n'));
+					yield* batch.map(async ticket => {
+						const task = pool.queue(w => w.exportTicket(ticket));
+						exportTasks[id].push(task);
+						return await task + '\n';
+					});
 					// Readable.from(AsyncGenerator) seems to be faster than pushing to a Readable with an empty `read()` function
 					// for (const ticket of batch) {
 					// 	pool
@@ -108,6 +158,57 @@ module.exports.get = fastify => ({
 			}
 		}
 
+		// V2
+		// const ticketsStream = Readable.from(ticketsGenerator());
+		// async function* ticketsGenerator() {
+		// 	try {
+		// 		let done = false;
+		// 		const take = 50;
+		// 		const findOptions = {
+		// 			include: {
+		// 				archivedChannels: true,
+		// 				archivedMessages: true,
+		// 				archivedRoles: true,
+		// 				archivedUsers: true,
+		// 				feedback: true,
+		// 				questionAnswers: true,
+		// 			},
+		// 			orderBy: { id: 'asc' },
+		// 			take,
+		// 			where: { guildId: id },
+		// 		};
+		// 		// create worker index array
+		// 		if (!exportTasks[id]) {
+		// 			exportTasks[id] = [];
+		// 		}
+		//
+		// 		do {
+		// 			const batch = await client.prisma.ticket.findMany(findOptions);
+		// 			if (batch.length < take) {
+		// 				done = true;
+		// 			} else {
+		// 				findOptions.skip = 1;
+		// 				findOptions.cursor = { id: batch[take - 1].id };
+		// 			}
+		// 			// ! map (parallel) not for...of (serial)
+		// 			// yield* batch.map(async ticket => (await pool.queue(w => w.exportTicket(ticket)) + '\n'));
+		// 			client.log.info(`Batch ${exportTasks[id].length}: queued`);
+		// 			const queuedPool = pool.queue(async w => w.exportTicketBatch(batch));
+		// 			exportTasks[id].push(queuedPool);
+		// 			queuedPool.then(result => {
+		// 				const qPool = queuedPool;
+		// 				const index = exportTasks[id].indexOf(qPool);
+		// 				client.log.info(`Batch ${index}: finished`);
+		// 				// Maybe remove task from list, but not for now, as we do that at the end
+		// 				// exportTasks[id].splice(index, 1);
+		// 			});
+		// 		} while (!done);
+		// 	} finally {
+		// 		yield* exportTasks[id];
+		// 		ticketsStream.push(null); // ! extremely important
+		// 	}
+		// }
+
 		const icon = await fetch(iconURL(guild));
 		archive.append(Readable.from(icon.body), { name: 'icon.png' });
 		archive.append(JSON.stringify(settings), { name: 'settings.json' });
@@ -115,12 +216,14 @@ module.exports.get = fastify => ({
 		archive.finalize(); // ! do not await
 
 		const cleanGuildName = guild.name.replace(/\W/g, '_').replace(/_+/g, '_');
-		const fileName = `tickets-${cleanGuildName}-${new Date().toISOString().slice(0, 10)}`;
+		const fileName = `tickets-${cleanGuildName}-${new Date().toISOString().slice(0, 10)}.zip`;
 
 		res
 			.type('application/zip')
 			.header('content-disposition', `attachment; filename="${fileName}"`)
-			.send(archive);
+			.send(archive)
+			// Release export lock on request closure or error
+			.then(() => releaseExport(id), () => releaseExport(id));
 	},
 	onRequest: [fastify.authenticate, fastify.isAdmin],
 });