+++ src/routes/api/admin/guilds/[guild]/panels.js
@@ -59,8 +59,7 @@ module.exports.post = fastify => ({
 		}
 
 		const embed = new EmbedBuilder()
-			.setColor(settings.primaryColour)
-			.setTitle(data.title);
+			.setColor(settings.primaryColour);
 
 		if (settings.footer) {
 			embed.setFooter({
@@ -69,6 +68,7 @@ module.exports.post = fastify => ({
 			});
 		}
 
+		if (data.title) embed.setTitle(data.title);
 		if (data.description) embed.setDescription(data.description);
 		if (data.image) embed.setImage(data.image);
 		if (data.thumbnail) embed.setThumbnail(data.thumbnail);