+++ src/routes/api/client.js
@@ -2,8 +2,12 @@
 
 const ms = require('ms');
 const pkg = require('../../../package.json');
-const { getAverageTimes } = require('../../lib/stats');
-const { quick } = require('../../lib/threads');
+const {
+	getAverageTimes, getAverageRating,
+} = require('../../lib/stats');
+const { pools } = require('../../lib/threads');
+
+const { stats } = pools;
 
 module.exports.get = () => ({
 	handler: async req => {
@@ -21,13 +25,14 @@ module.exports.get = () => ({
 				users,
 			] = await Promise.all([
 				client.prisma.category.count(),
-				quick('stats', w => w.sum(client.guilds.cache.map(g => g.memberCount))),
+				stats.queue(w => w.sum(client.guilds.cache.map(g => g.memberCount))),
 				client.prisma.tag.count(),
 				client.prisma.ticket.count(),
 				client.prisma.ticket.findMany({
 					select: {
 						closedAt: true,
 						createdAt: true,
+						feedback: { select: { rating: true } },
 						firstResponseAt: true,
 					},
 					where: {
@@ -44,6 +49,7 @@ module.exports.get = () => ({
 				avgResolutionTime,
 				avgResponseTime,
 			} = await getAverageTimes(closedTickets);
+			const avgRating = await getAverageRating(closedTickets);
 
 			cached = {
 				avatar: client.user.avatarURL(),
@@ -53,6 +59,7 @@ module.exports.get = () => ({
 				stats: {
 					activatedUsers: users._count,
 					archivedMessages: users._sum.messageCount,
+					avgRating: avgRating.toFixed(1),
 					avgResolutionTime: ms(avgResolutionTime),
 					avgResponseTime: ms(avgResponseTime),
 					categories,